+++ release/ngx-ui.module.js
@@ -11,7 +11,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { InjectionService, IconRegisteryService } from './services';
 import { DirectivesModule } from './directives';
 import { PipesModule } from './pipes';
-import { CalendarModule, CodeEditorModule, LoadingModule, DrawerModule, DropdownModule, ButtonModule, InputModule, SectionModule, SliderModule, TabsModule, ToolbarModule, TooltipModule, OverlayModule, DialogModule, OverlayService, DialogService, DrawerService, TooltipService, ToggleModule, DateTimeModule, CheckboxModule, NotificationModule, NotificationService, SelectModule, IconModule, LoadingService, TreeModule, SplitModule, HotkeysModule, NagModule } from './components';
+import { CalendarModule, CodeEditorModule, LoadingModule, DrawerModule, DropdownModule, ButtonModule, LongPressButtonModule, InputModule, SectionModule, SliderModule, TabsModule, ToolbarModule, TooltipModule, OverlayModule, DialogModule, OverlayService, DialogService, DrawerService, TooltipService, ToggleModule, DateTimeModule, CheckboxModule, NotificationModule, NotificationService, SelectModule, IconModule, LoadingService, TreeModule, SplitModule, HotkeysModule, NagModule } from './components';
 /**
  * Exported Modules
  * @type {Array}
@@ -23,7 +23,8 @@ var modules = [
     ToolbarModule, TooltipModule, CommonModule, FormsModule,
     OverlayModule, DialogModule, ToggleModule, DateTimeModule,
     CheckboxModule, NotificationModule, PipesModule, SelectModule,
-    IconModule, LoadingModule, TreeModule, SplitModule, HotkeysModule, NagModule
+    IconModule, LoadingModule, TreeModule, SplitModule, HotkeysModule, NagModule,
+    LongPressButtonModule
 ];
 var NgxUIModule = /** @class */ (function () {
     function NgxUIModule() {